DAT-4115 move navigation items to model class
